feat: implemented userId check after authmiddleware on user functions

+++ src/controllers/wishlistController.ts
@@ -2,12 +2,23 @@ import { Response, Request, NextFunction } from "express";
 import pool from "../config/db";
 import type { Wishlist} from "../types";
 
-async function createWishlist(req: Request, res: Response, _next: NextFunction): Promise<void>{
+interface AuthenticatedRequest extends Request {
+    user?: { id: string; email: string }; 
+}
+
+async function createWishlist(req: AuthenticatedRequest , res: Response, _next: NextFunction): Promise<void>{
 
     const client = await pool.connect();
 
+     if (!req.user) {
+            res.status(401).json({ msg: "Utente non autenticato" });
+            return;
+        }
+    
+        const userId = req.user.id; 
+
     const wishlist: Wishlist = req.body;
-    const { name, gifts, userId } = wishlist;
+    const { name, gifts} = wishlist;
 
     if(!userId){
         res.status(400).json({msg: "Errore Creazione Wishlist: Errore attribuzione proprietario"});
@@ -67,6 +78,8 @@ async function createWishlist(req: Request, res: Response, _next: NextFunction):
             console.error(error);
             res.status(500).json({msg: "Database Error: Creazione Wishlist non avvenuta"})
             return;
+    }finally {
+    client.release(); 
     }
    
 
@@ -105,8 +118,14 @@ async function getPublicWishlist(req: Request, res: Response): Promise<void> {
     }
 }
 
-async function deleteWishlist(req: Request, res: Response): Promise<void>{
-    const {id} = req.params;
+async function deleteWishlist(req: AuthenticatedRequest , res: Response): Promise<void>{
+     if (!req.user) {
+            res.status(401).json({ msg: "Utente non autenticato" });
+            return;
+        }
+
+        const userId = req.user.id; 
+        const {id} = req.params;
 
     try {
         if(!id){
@@ -121,7 +140,7 @@ async function deleteWishlist(req: Request, res: Response): Promise<void>{
             return;
         }
 
-        const removedWishlist = await pool.query('DELETE FROM wishlists WHERE id = $1 RETURNING id, name', [parseId]);
+        const removedWishlist = await pool.query('DELETE FROM wishlists WHERE id = $1 AND user_id = $2 RETURNING id, name', [parseId, userId]);
 
         if(!removedWishlist.rowCount || removedWishlist.rowCount === 0){
             res.status(404).json({msg: "Nessuna wishlist presente con l'ID inserito, la rimozione non è avvenuta"});
@@ -137,8 +156,48 @@ async function deleteWishlist(req: Request, res: Response): Promise<void>{
     }
 }
 
-async function updateWishlist(req: Request, res: Response): Promise<void>{
+async function updateWishlist(req: AuthenticatedRequest , res: Response): Promise<void>{
+    try {
+         if (!req.user) {
+            res.status(401).json({ msg: "Utente non autenticato" });
+            return;
+        }
+        // L'ID dell'utente lo prendiamo dal TOKEN, non dai parametri!
+        const userId = req.user.id; 
+        
+    } catch (error) {
+        
+    }
+}
 
+async function getMyWishlists(req: AuthenticatedRequest , res: Response): Promise<void>{
+    try {
+        if (!req.user) {
+            res.status(401).json({ msg: "Utente non autenticato" });
+            return;
+        }
+        // L'ID dell'utente lo prendiamo dal TOKEN, non dai parametri!
+        const userId = req.user.id; 
+
+        // Recuperiamo le wishlist create dall'utente
+        const myWishlistsResult = await pool.query(
+            `SELECT * from wishlists where user_id = $1`, [userId]
+        );
+
+        // Recuperiamo le preferite (assumendo che tu abbia un array o una tabella pivot)
+        const favoritesResult  = await pool.query(
+            `SELECT w.* FROM wishlists w
+                 JOIN favorite_whishlists f ON w.id = f.wishlist_id
+                 WHERE f.user_id = $1`, [userId]
+        );
+
+        res.json({
+            wishlists: myWishlistsResult.rows,
+            favorites: favoritesResult.rows
+        });
+    } catch (error) {
+        res.status(500).json({ msg: "Errore nel recupero delle liste" });
+    }
 }
 
 
@@ -146,5 +205,6 @@ export {
     createWishlist,
     getPublicWishlist,
     deleteWishlist,
-    updateWishlist
+    updateWishlist,
+    getMyWishlists
 }
